1) Did the constraints correctly for all bricks. Should render correctly in all iPhones (4,5,6,6+) In portrait mode.

+++ Bouncy Ball Attempt 2/ViewController.m
@@ -168,8 +168,9 @@ BOOL brick;
             pos.y = -(pos.y);
             brick.hidden = true;
           //  NSLog(@"value bla: %d", brick.hidden);
-            NSLog(@"Hidden brick is: %d",brick.tag);
+            NSLog(@"Hidden brick is: %ld",(long)brick.tag);
             [self.audioController playSystemSound];
+
         }
     }
